TauValidationNTupleProd: Add pt and eta cuts on objects for tree:
Skim jobs on samples with large number of vertices fail on the grid
due to too large RAM usage.
Since a large number of information is stored for each pfJet candidate
this could be the reason if there are too many candidates in one event.
--> Introduce a cut on pt and eta on skimmer level.
The cuts are weaker than the ones applied offline.